Student Details

+++ src/main/java/com/ust/training/studentdata/service/StudentService.java
@@ -9,6 +9,7 @@ import com.ust.training.studentdata.model.Student;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.stereotype.Service;
@@ -18,10 +19,7 @@ import com.ust.training.studentdata.Exception.StudentServiceException;
 import com.ust.training.studentdata.common.StudentDTO;
 import com.ust.training.studentdata.repo.StudentRepo;
 
-import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
-
-
 
 /**
  * 
@@ -29,8 +27,6 @@ import reactor.core.publisher.Mono;
  *
  */
 @Service
-
-@Slf4j
 public class StudentService {
   @Autowired
   private StudentRepo repository;
@@ -41,26 +37,22 @@ public class StudentService {
    * Save student details
    * 
    * @param studentDTO
-   * @return
+   * @return student obj
    */
-  public String saveStudentDetails(StudentDTO studentDTO) {
+  public Student saveStudentDetails(StudentDTO studentDTO) {
 
     Student details = new Student();
     details.setId(studentDTO.getId());
     details.setFirstName(studentDTO.getFirstName());
     details.setLastName(studentDTO.getLastName());
     details.setAddress(studentDTO.getAddress());
-
     log.debug("Saving studentDetails: {}", details);
-
     // Save the User class to Azure CosmosDB database.
     Mono<Student> saveStudent = repository.save(details);
-
     // Nothing happens until we subscribe to these Monos.
     Student savedStudent = saveStudent.block();
     log.debug("Saved StudentDetails");
-
-    return "Save Success";
+    return savedStudent;
   }
 
   /***
@@ -70,13 +62,19 @@ public class StudentService {
    * @return student obj
    */
 
-  public Student getStudentDetails(String id) {
-    log.debug("inside get method");
+  public StudentDTO getStudentDetails(String id) {
+    StudentDTO studentDTO = new StudentDTO();
+    log.debug("Begining of get method");
     try {
-      Mono<Student> findByIdStudent = repository.findById(id);
-      Student findStudentById = findByIdStudent.block();
+      Student findByIdStudent = repository.findById(id).block();
+      // Student findStudentById = findByIdStudent.block();
+      if (null == findByIdStudent) {
 
-      return findStudentById;
+        return null;
+      }
+      log.debug("Ending of get method");
+      BeanUtils.copyProperties(findByIdStudent, studentDTO);
+      return studentDTO;
 
     } catch (Exception e) {
       log.error("Exception:", e);
@@ -93,17 +91,22 @@ public class StudentService {
    * @return String
    */
   public String deleteStudentDetails(String id) {
-    log.debug("inside delete method");
+    log.debug("Begining of delete method");
     try {
-      Mono<Student> findByIdStudent = repository.findById(id);
-      Student findStudentById = findByIdStudent.block();
-      repository.delete(findStudentById).block();
-
+      Student findStudentById = repository.findById(id).block();
+      // Student findStudentById = findByIdStudent.block();
+      if (null != findStudentById) {
+        repository.delete(findStudentById).block();
+        log.debug("Ending of delete method");
+        return "Student deleted";
+      }
     } catch (Exception e) {
       log.error("Exception:", e);
       throw new StudentServiceException("Exception in DeleteStudentDetails", e);
     }
-    return "Deleted";
+    return "Student not deleted";
+
+
   }
 
   /***
@@ -115,28 +118,23 @@ public class StudentService {
 
   public Student updateStudent(StudentDTO studentDTO) {
     Student savedStudent = null;
+    log.debug("inside update student");
     try {
-      log.debug("inside update student");
+
       Student studentDetails = new Student();
       Mono<Student> findByIdStudent = repository.findById(studentDTO.getId());
       Student student = findByIdStudent.block();
-
-      repository.delete(student).block();
-
-      studentDetails.setId(studentDTO.getId());
-
-      studentDetails.setFirstName(studentDTO.getFirstName());
-
-      studentDetails.setLastName(studentDTO.getLastName());
-
-      studentDetails.setAddress(studentDTO.getAddress());
-
-      Mono<Student> saveStudent = repository.save(student);
-
-      savedStudent = saveStudent.block();
-
+      if (null != studentDetails) {
+        repository.delete(student).block();
+        studentDetails.setId(studentDTO.getId());
+        studentDetails.setFirstName(studentDTO.getFirstName());
+        studentDetails.setLastName(studentDTO.getLastName());
+        studentDetails.setAddress(studentDTO.getAddress());
+        Mono<Student> saveStudent = repository.save(student);
+        savedStudent = saveStudent.block();
+      }
       return savedStudent;
-      
+
     } catch (Exception exception) {
       log.error("Exception:", exception);
       throw new StudentServiceException("Exception in Saving a Student", exception);
